perf: improve l0 seclection (#24417)
